requirements.txt
Pillow
Flask
colorama
gunicorn
ih
